requirements.txt
flask-restful
pymongo[srv]
flask-cors
pandas